Be more defensive.
We eliminated tripping a warning over an edge case. If a game cubes of
a certain colour are not shown, we'll still calculate the right value,
but we will be getting an "uninitialized" warning.

The input doesn't contain such cases, but they aren't ruled out in the
description, so we must handle this case.